Highlight available comments on overview if query enabled
The availability of comments on the test result page is shown on the test
result overview page next to the status icon and failed modules if available
and only if selected with the query parameter "show_comment=1".

The icon only appears if comments are available. This way it should be much
easier for reviewers of test runs to exchange and see if an issue has already
been noticed by someone else.

Be aware that so far the comments are per test result not per test case or
scenario.

Related issue: https://progress.opensuse.org/issues/10212